⚙️ Update Python versions in package_testing.yml
Removed Python 3.7 version and added Python 3.11.312 versions in the package_testing.yml file. This change addresses the error encountered during installation, where the requirement for flake8==6.0.0 could not be satisfied with the available distributions.

By updating the Python versions, we ensure compatibility with the required dependencies and enable successful installation and testing.