🔧 Fix Python version and ✨ Update Torch versions in ci.yml
Fixed the Python version in the ci.yml file to resolve the error where Python 3.12 was not found in the local cache. Replaced Python 3.12 with Python 3.11 to align with the available versions. Additionally, updated the Torch versions in the ci.yml file by adding Torch 2.0.0 and 2.0.1 for testing. The updated list of Torch versions now includes: [1.10.2, 1.11.0, 1.12.0, 1.13.0, 2.0.0, 2.0.1].

These changes improve the testing configuration and ensure compatibility with a wider range of Torch releases.